fix lints, move secrets reader to different package
