+++ system-tests/lib/cre/don/don.go
@@ -1,7 +1,6 @@
 package don
 
 import (
-	"fmt"
 	"slices"
 	"strconv"
 
@@ -13,7 +12,6 @@ import (
 	"github.com/smartcontractkit/chainlink/system-tests/lib/cre/don/node"
 	"github.com/smartcontractkit/chainlink/system-tests/lib/cre/flags"
 	cretypes "github.com/smartcontractkit/chainlink/system-tests/lib/cre/types"
-	"github.com/smartcontractkit/chainlink/system-tests/lib/crypto"
 	"github.com/smartcontractkit/chainlink/system-tests/lib/infra"
 	"github.com/smartcontractkit/chainlink/system-tests/lib/types"
 )
@@ -157,129 +155,6 @@ func BuildTopology(nodeSetInput []*cretypes.CapabilitiesAwareNodeSet, infraInput
 	topology.WorkflowDONID = maybeID.ID
 
 	return topology, nil
-}
-
-func AddKeysToTopology(topology *cretypes.Topology, keys *cretypes.GenerateKeysOutput) (*cretypes.Topology, error) {
-	if topology == nil {
-		return nil, errors.New("topology is nil")
-	}
-
-	if keys == nil {
-		return nil, errors.New("keys is nil")
-	}
-
-	if len(keys.P2PKeys) != len(topology.DonsMetadata) {
-		return nil, fmt.Errorf("number of P2P keys does not match the number of DONs. Expected %d, got %d", len(topology.DonsMetadata), len(keys.P2PKeys))
-	}
-
-	if len(keys.EVMKeys) != len(topology.DonsMetadata) {
-		return nil, fmt.Errorf("number of EVM keys does not match the number of DONs. Expected %d, got %d", len(topology.DonsMetadata), len(keys.EVMKeys))
-	}
-
-	for _, donMetadata := range topology.DonsMetadata {
-		if p2pKeys, ok := keys.P2PKeys[donMetadata.ID]; ok {
-			if len(p2pKeys.PeerIDs) != len(donMetadata.NodesMetadata) {
-				return nil, fmt.Errorf("number of P2P keys for DON %d does not match the number of nodes. Expected %d, got %d", donMetadata.ID, len(donMetadata.NodesMetadata), len(p2pKeys.PeerIDs))
-			}
-			for idx, nodeMetadata := range donMetadata.NodesMetadata {
-				nodeMetadata.Labels = append(nodeMetadata.Labels, &cretypes.Label{
-					Key:   node.NodeP2PIDKey,
-					Value: p2pKeys.PeerIDs[idx],
-				})
-			}
-		} else {
-			return nil, fmt.Errorf("no P2P keys found for DON %d", donMetadata.ID)
-		}
-
-		if chainIDsToEVMKeys, ok := keys.EVMKeys[donMetadata.ID]; ok {
-			// First, verify that all chain IDs have the same EVM keys for each node
-			// This is a limitation of our current testing SDK implementation, because we only have 1 label for ETH address for each node
-			// If in the future we need to support multiple chain IDs, we will need to change this and prefix the label with the chain ID
-			// For now let's just make sure that all the chain IDs have the same EVM keys for each node to avoid hard to debug issues
-			var firstChainID int
-			var firstEVMKeys *types.EVMKeys
-			for chainID, evmKeys := range chainIDsToEVMKeys {
-				if firstEVMKeys == nil {
-					firstChainID = chainID
-					firstEVMKeys = evmKeys
-					continue
-				}
-				if len(evmKeys.PublicAddresses) != len(firstEVMKeys.PublicAddresses) {
-					return nil, fmt.Errorf("number of EVM keys for DON %d differs between chain IDs %d and %d", donMetadata.ID, firstChainID, chainID)
-				}
-				for i := range evmKeys.PublicAddresses {
-					if evmKeys.PublicAddresses[i] != firstEVMKeys.PublicAddresses[i] {
-						return nil, fmt.Errorf("EVM public address mismatch for DON %d, node %d between chain IDs %d and %d", donMetadata.ID, i, firstChainID, chainID)
-					}
-				}
-			}
-
-			// Now add the EVM addresses to the node metadata
-			for chainID, evmKeys := range chainIDsToEVMKeys {
-				if len(evmKeys.PublicAddresses) != len(donMetadata.NodesMetadata) {
-					return nil, fmt.Errorf("number of EVM keys for DON %d and chain ID %d does not match the number of nodes. Expected %d, got %d", donMetadata.ID, chainID, len(donMetadata.NodesMetadata), len(evmKeys.PublicAddresses))
-				}
-				for idx, nodeMetadata := range donMetadata.NodesMetadata {
-					nodeMetadata.Labels = append(nodeMetadata.Labels, &cretypes.Label{
-						Key:   node.EthAddressKey,
-						Value: evmKeys.PublicAddresses[idx].Hex(),
-					})
-				}
-				// Use first ETH address for the DON metadata, because all of them are the same
-				break
-			}
-		} else {
-			return nil, fmt.Errorf("no EVM keys found for DON %d", donMetadata.ID)
-		}
-	}
-
-	return topology, nil
-}
-
-func GenereteKeys(input *cretypes.GenerateKeysInput) (*cretypes.GenerateKeysOutput, error) {
-	if input == nil {
-		return nil, errors.New("input is nil")
-	}
-
-	if err := input.Validate(); err != nil {
-		return nil, errors.Wrap(err, "input validation failed")
-	}
-
-	if input.Out != nil {
-		return input.Out, nil
-	}
-
-	output := &cretypes.GenerateKeysOutput{
-		EVMKeys: make(cretypes.DonsToEVMKeys),
-		P2PKeys: make(cretypes.DonsToP2PKeys),
-	}
-
-	for _, donMetadata := range input.Topology.DonsMetadata {
-		if input.GenerateP2PKeys {
-			p2pKeys, err := crypto.GenerateP2PKeys(input.Password, len(donMetadata.NodesMetadata))
-			if err != nil {
-				return nil, errors.Wrap(err, "failed to generate P2P keys")
-			}
-			output.P2PKeys[donMetadata.ID] = p2pKeys
-		}
-
-		if len(input.GenerateEVMKeysForChainIDs) > 0 {
-			evmKeys, err := crypto.GenerateEVMKeys(input.Password, len(donMetadata.NodesMetadata))
-			if err != nil {
-				return nil, errors.Wrap(err, "failed to generate EVM keys")
-			}
-
-			// use the same EVM keys for all the chain IDs
-			for _, chainID := range input.GenerateEVMKeysForChainIDs {
-				if _, ok := output.EVMKeys[donMetadata.ID]; !ok {
-					output.EVMKeys[donMetadata.ID] = make(cretypes.ChainIDToEVMKeys)
-				}
-				output.EVMKeys[donMetadata.ID][chainID] = evmKeys
-			}
-		}
-	}
-
-	return output, nil
 }
 
 func NodeNeedsGateway(nodeFlags []cretypes.CapabilityFlag) bool {
